+++ packages/manifest/src/index.ts
@@ -12,10 +12,12 @@ export const name = 'True - False';
 
 // Function which inits element state (data property on the Content Element
 // entity)
-// e.g. for simple counter component:
 export const initState: DataInitializer = (): ElementData => ({
-  question: '',
   correct: null,
+  embeds: {},
+  question: [],
+  hint: '',
+  feedback: {},
 });
 
 // Can be loaded from package.json
@@ -30,18 +32,15 @@ const ui = {
   forceFullWidth: true,
 };
 
-export const mocks = {
-  displayContexts: [{ name: 'No selection', data: {} }],
-};
-
 const manifest: ElementManifest = {
   type,
   version: '1.0',
   name,
+  isComposite: true,
+  isQuestion: true,
   ssr: false,
   initState,
   ui,
-  mocks,
 };
 
 export default manifest;